+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [0.3.1] - 2022-04-23
+
+- Show all planets.
+- Tweak zoom boundaries for the Solar System.
+- Better colors for planets.
+
 ## [0.3.0] - 2022-04-23
 
 - Started to track this CHANGELOG.